src/main/java/com/thoughtworks/capability/gtb/restfulapidesign/service/StudentService.java
package com.thoughtworks.capability.gtb.restfulapidesign.service;

import com.thoughtworks.capability.gtb.restfulapidesign.dto.Student;
import com.thoughtworks.capability.gtb.restfulapidesign.repository.StudentRepository;
import org.springframework.stereotype.Service;

@Service
public class StudentService {

    private final StudentRepository studentRepository;

    public StudentService(StudentRepository studentRepository) {
        this.studentRepository = studentRepository;
    }

    public Student createStudent(Student student) {
        student.setId(studentRepository.getMaxStudentId());
        studentRepository.setMaxStudentId(student.getId() + 1);
        studentRepository.save(student);
        return student;
    }
}
